381. Insert Delete GetRandom O(1) - Duplicates allowed
